Create vehicleDetectionYOLOv2.py
This method of vehicle detection is one of the best. But the sad part is, in this approach, we have to use older versions of all the libraries. Basically use all the libraries of python2, which has been deprecated. Furthermore, after solving all the inherent dependencies and deprecations, while running predictions on the model and preprocessing, we get errors that we cant initialize the predictions from the first pass. Even on changing it to a tensor object, it asks for a Model instance in the predict method. But such is not mentioned in the documentation of tf.keras.Model.predict(x). This would be considered a work in develpment.